requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
flask-paginate==0.5.4
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.2
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
